[Feat] 배송 서비스 구현
- 배송시작 이전까지 배송 주소 변경 가능
- 주문서비스에서 사용하는 배송 상태 응답 객체 추가

+++ src/main/java/com/barogo/api/domain/delivery/repository/DeliveryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryRepository extends JpaRepository<Delivery, Integer>, DeliveryQueryRepository {
+public interface DeliveryRepository extends JpaRepository<Delivery, Long>, DeliveryQueryRepository {
 }
